requirements.txt
gradio==4.19.2
sentence-transformers==2.5.1
faiss-cpu==1.7.4
numpy==1.26.4
requests==2.31.0 